requirements.txt
cryptg
python-dotenv==0.19.2
Telethon==1.24.0
Gino==1.0.1
SQLAlchemy~=1.3.24
python-slugify
boto3